requirements/dev.txt
-r base.txt

ipython==8.13
psycopg2==2.9.7
python-environ==0.4.54
pillow==10.0.0
mypy==1.6.1
django-stubs==1.2.0
whitenoise==6.5.0
channels==4.0.0
daphne==4.0.0 
channels-redis==4.1.0 